third_party/blink/web_tests/html5lib/generated/run-webkit01-write-expected.txt

CONSOLE MESSAGE: PASS
CONSOLE MESSAGE: FOO<span>BAR</span>BAZ
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select>. Please add the missing </select> end tag.
../resources/webkit01.dat: PASS
